refactor(theme2): consolidate theme2 modules into a single module
