+++ app-ui-catalog/build.gradle.kts
@@ -31,8 +31,7 @@ dependencies {
     implementation(projects.core.ui.compose.designsystem)
     implementation(projects.core.ui.legacy.designsystem)
 
-    implementation(projects.core.ui.compose.theme2.thunderbird)
-    implementation(projects.core.ui.compose.theme2.k9mail)
+    implementation(projects.core.ui.compose.theme2)
 
     implementation(libs.androidx.compose.material3)
     implementation(libs.androidx.compose.material.icons.extended)
