--- core/ui/compose/theme2/common/build.gradle.kts
@@ -1,25 +1,0 @@
-plugins {
-    id(ThunderbirdPlugins.Library.kmpCompose)
-}
-
-kotlin {
-    android {
-        namespace = "net.thunderbird.core.ui.compose.theme2"
-        androidResources.enable = true
-    }
-
-    sourceSets {
-        androidMain.dependencies {
-            implementation(libs.android.material)
-        }
-
-        commonMain.dependencies {
-            implementation(libs.jetbrains.compose.material3)
-        }
-    }
-}
-
-codeCoverage {
-    branchCoverage = 0
-    lineCoverage = 0
-}
