--- core/ui/compose/theme2/k9mail/README.md
@@ -1,13 +1,0 @@
-## Core - UI - Compose - Theme2 - K9Mail
-
-This provides the `K9MailTheme2` composable, that's setting up the `MainTheme` with K-9 Mail specific colors, typography, shapes, elevations, sizes, spacings and images.
-
-```kotlin
-@Composable
-fun MyComposable() {
-    K9MailTheme2 {
-        // Your app content
-    }
-}
-```
-
