+++ core/ui/compose/theme2/src/commonMain/composeResources/drawable/core_ui_theme2_thunderbird_logo.xml
