+++ core/ui/compose/theme2/src/commonMain/kotlin/net/thunderbird/core/ui/compose/theme2/ThemeShapes.kt
