+++ feature/migration/qrcode/build.gradle.kts
@@ -30,7 +30,6 @@ dependencies {
 
     testImplementation(projects.core.logging.testing)
     testImplementation(projects.core.ui.compose.testing)
-    testImplementation(projects.core.ui.compose.theme2.k9mail)
 }
 
 codeCoverage {
