+++ feature/notification/api/build.gradle.kts
@@ -22,7 +22,7 @@ kotlin {
         }
         androidMain.dependencies {
             implementation(projects.core.ui.compose.designsystem)
-            implementation(projects.core.ui.compose.theme2.common)
+            implementation(projects.core.ui.compose.theme2)
         }
         androidHostTest.dependencies {
             implementation(projects.core.ui.compose.testing)
